requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cleo==0.8.1
click==8.2.1
clikit==0.6.2
crashtest==0.3.1
dnspython==2.7.0
email_validator==2.2.0
Faker==13.16.0
fastapi==0.115.12
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
masonite-orm==3.0.0
mdurl==0.1.2
pastel==0.2.1
pendulum==3.1.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pylev==1.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
